chore(TemperatureData): move climate.ts to features/climate/
