+++ frontend/src/features/climate/components/TemperatureChart.tsx
@@ -11,7 +11,7 @@ import {
 import { useEffect, useState } from "react";
 import { useQuery } from "@tanstack/react-query";
 import { fetchTemperatureData } from "@/features/climate/api/climateApi";
-import type { TemperatureData } from "@/types/models/climate";
+import type { TemperatureData } from "@/features/climate/types/climate";
 import { Loading, SelectBox } from "@/components/common";
 import { TemperatureChartDescription } from "@/features/climate/components/TemperatureChartDescription";
